+++ noodles-tabix/src/lib.rs
@@ -19,17 +19,23 @@
 //! ## Read a tabix file
 //!
 //! ```no_run
-//! # use std::io;
 //! use noodles_tabix as tabix;
-//! let index = tabix::read("sample.vcf.gz.tbi")?;
-//! # Ok::<(), io::Error>(())
+//! let index = tabix::fs::read("sample.vcf.gz.tbi")?;
+//! # Ok::<(), std::io::Error>(())
 //! ```
 
 #[cfg(feature = "async")]
 pub mod r#async;
 
+pub mod fs;
 pub mod index;
 pub mod io;
+
+#[deprecated(since = "0.48.0", note = "Use `tabix::fs::read` instead.")]
+pub use self::fs::read;
+
+#[deprecated(since = "0.48.0", note = "Use `tabix::fs::write` instead.")]
+pub use self::fs::write;
 
 #[deprecated(since = "0.45.0", note = "Use `noodles_tabix::io::Reader` instead.")]
 pub use self::io::Reader;
@@ -51,55 +57,9 @@ pub use self::r#async::Reader as AsyncReader;
 )]
 pub use self::r#async::Writer as AsyncWriter;
 
-use std::{fs::File, path::Path};
-
 use noodles_csi::binning_index::{self, index::reference_sequence::index::LinearIndex};
 
 const MAGIC_NUMBER: [u8; 4] = *b"TBI\x01";
 
 /// A tabix index.
 pub type Index = binning_index::Index<LinearIndex>;
-
-/// Reads the entire contents of a tabix index.
-///
-/// This is a convenience function and is equivalent to opening the file at the given path and
-/// reading the index.
-///
-/// # Examples
-///
-/// ```no_run
-/// use noodles_tabix as tabix;
-/// let index = tabix::read("sample.vcf.gz.tbi")?;
-/// # Ok::<(), std::io::Error>(())
-/// ```
-pub fn read<P>(src: P) -> std::io::Result<Index>
-where
-    P: AsRef<Path>,
-{
-    let mut reader = File::open(src).map(Reader::new)?;
-    reader.read_index()
-}
-
-/// Writes a tabix index to a file.
-///
-/// This is a convenience function and is equivalent to creating a file at the given path and
-/// writing the index.
-///
-/// # Examples
-///
-/// ```no_run
-/// use noodles_csi::binning_index::index::Header;
-/// use noodles_tabix as tabix;
-///
-/// let index = tabix::Index::builder().set_header(Header::default()).build();
-/// tabix::write("sample.vcf.gz.tbi", &index)?;
-/// # Ok::<(), std::io::Error>(())
-/// ```
-pub fn write<P>(dst: P, index: &Index) -> std::io::Result<()>
-where
-    P: AsRef<Path>,
-{
-    let mut writer = File::create(dst).map(Writer::new)?;
-    writer.write_index(index)?;
-    Ok(())
-}